Repaired remove functionality, added add site on popup Wow, there were a lot of things that I luckily didn't break on my last tests. Granted, they were very limited test cases, but still.
This commit possibly introduces an issue with a first element being '', 
but siteBlocked has been trained to ignore this entry. I get it - this 
is pretty straightforward text parsing at this point - but even still 
that blank element will be a quirk. At least the code is built to ignore 
it, and if the Options interface isn't ready for it yet, it will be 
soon.